contrib/Fun_TaskGen_Projects/web_browser_implementation/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
attrs==23.2.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
dill==0.3.8
distro==1.9.0
exceptiongroup==1.2.0
google-api-core==2.18.0
google-auth==2.29.0
google-cloud-vision==3.7.2
googleapis-common-protos==1.63.0
grpcio==1.62.2
grpcio-status==1.62.2
h11==0.14.0
helium==5.0.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
MouseInfo==0.1.3
numpy==1.26.4
openai==1.3.6
outcome==1.3.0.post0
pillow==10.2.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAutoGUI==0.9.54
pydantic==2.6.4
pydantic_core==2.16.3
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyobjc-core==10.1
pyobjc-framework-Cocoa==10.1
pyobjc-framework-Quartz==10.1
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
python-dotenv==1.0.1
pytweening==1.2.0
requests==2.31.0
rsa==4.9
rubicon-objc==0.4.7
selenium==4.18.1
sniffio==1.3.1
sortedcontainers==2.4.0
taskgen-ai==1.2.0
tqdm==4.66.2
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.10.0
urllib3==2.2.1
wsproto==1.2.0
